requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
streamlit==1.22.0
pandas==1.5.3
numpy==1.24.2
yfinance==0.2.18
plotly==5.14.1
pyyaml==6.0
networkx==3.1
aiofiles==23.1.0
requests==2.31.0
python-multipart==0.0.6
types-aiofiles==23.1.0.1
types-requests==2.31.0.1